requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
Pillow>=10.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
wordcloud>=1.9.0
python-dotenv>=1.0.0
